[level 0] Title: 문자 개수 세기, Time: 0.20 ms, Memory: 31.2 MB -BaekjoonHub

+++ 프로그래머스/0/181902. 문자 개수 세기/문자 개수 세기.cs
@@ -5,7 +5,7 @@ public class Solution {
         int[] answer = new int[52];
         
         foreach(char word in my_string){
-            int num = (int)word>64 && (int)word<91 ? (int)word-65 : (int)word-71;
+            int num = word>64 && word<91 ? word-65 : word-71;
             answer[num]++;
         }
